MatcherExample/src/main/java/com/epam/Car.java
package com.epam;

public class Car  {
    private String name;
    private double engineVolume;
    private int yearProduction;

    public Car (String aName, double aEngineVolume, int aYearProduction) {
        this.name = aName;
        this.engineVolume = aEngineVolume;
        this.yearProduction = aYearProduction;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getEngineVolume() {
        return engineVolume;
    }

    public void setEngineVolume(double engineVolume) {
        this.engineVolume = engineVolume;
    }

    public int getYearProduction() {
        return yearProduction;
    }

    public void setYearProduction(int yearProduction) {
        this.yearProduction = yearProduction;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (!(o instanceof Car)) return false;

        Car car = (Car) o;

        if (Double.compare(car.engineVolume, engineVolume) != 0) return false;
        if (yearProduction != car.yearProduction) return false;
        return name != null ? name.equals(car.name) : car.name == null;
    }

    @Override
    public int hashCode() {
        int result;
        long temp;
        result = name != null ? name.hashCode() : 0;
        temp = Double.doubleToLongBits(engineVolume);
        result = 31 * result + (int) (temp ^ (temp >>> 32));
        result = 31 * result + yearProduction;
        return result;
    }

    @Override
    public String toString() {
        return "Car{" +
                "name='" + name + '\'' +
                ", engineVolume=" + engineVolume +
                ", yearProduction=" + yearProduction +
                '}';
    }
}
